% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codemirror_addons.R
\name{cm_foldcode}
\alias{cm_foldcode}
\title{Code Folding for CodeMirror}
\usage{
cm_foldcode(cm_widget = NULL, fold_mode = NULL)
}
\arguments{
\item{cm_widget}{\code{codemirror} htmlwidget}

\item{fold_mode}{\code{character} for the type of folding.  Valid options
are \code{"brace-fold"}, \code{"comment-fold"}, \code{"indent-fold"},
\code{"markdown-fold"}, and \code{"xml-fold"}.}
}
\value{
\code{codemirror} htmlwidget
}
\description{
Add code folding for your \code{codemirror} htmlwidget.  This is designed
  for easy piping with \code{magrittr} or \code{pipeR}.
}

