% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codemirror.R
\name{codemirror}
\alias{codemirror}
\title{<Add Title...>}
\usage{
codemirror(doc = "", mode = NULL, theme = NULL, addon = NULL, ...,
  width = NULL, height = NULL, elementId = NULL)
}
\description{
<Add Description>
}

